% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_goals.R
\name{list_sdg_goals}
\alias{list_sdg_goals}
\title{Les 17 Objectifs de developpement durable (ODD)}
\usage{
list_sdg_goals()
}
\value{
Un \code{data.frame} avec les colonnes \code{goal} et \code{title}.
}
\description{
Retourne un data.frame des 17 objectifs, avec leur numero et leur
titre officiel (traduction francaise des Nations Unies).
}
